# go-web-template/docs/swagger.yaml
basePath: /
definitions:
  model.User:
    properties:
      address:
        type: string
      age:
        maximum: 120
        minimum: 0
        type: integer
      email:
        type: string
      gender:
        enum:
        - 0
        - 1
        type: integer
      id:
        type: integer
      name:
        type: string
    required:
    - email
    - gender
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  title: go-web-template API
  version: "1.0"
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: name
        in: query
        name: name
        type: string
      - description: gender
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户列表
      tags:
      - 用户
    post:
      consumes:
      - application/json
      description: 新增一个用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建用户
      tags:
      - 用户
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除一个用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除用户
      tags:
      - 用户
    get:
      consumes:
      - application/json
      description: 根据ID获取一个用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取用户
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 根据ID更新一个用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新用户
      tags:
      - 用户
  /api/users/all:
    get:
      consumes:
      - application/json
      description: 获取所有用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取所有用户
      tags:
      - 用户
swagger: "2.0"